4.7 Registering Multiple Implementations of an Interface
Pay special attention to the file BookingRuleProcessor.cs and the changes on Startup.cs

This is an use case of registering multiple implementations of an interface. In a previous section we learned that the last implementation to be registered is the one used; but in this example all are used through IEnumerable<ICourtBookingRule>

This setup allows to add new rules without changing the implementation of the others, the only file that would need to be modified is Startup.cs to register the new implementation.