5.4.2 Action Injection
In this example, the Courts Controller uses 2 services:

_courtBookingService
_courtMaintenanceService

The latter is only used by one of the actions of the controller, to reduce memory and GC use, we can inject the service only for the specific action that uses it.

This is a symptom that a new controller for court maintenance should be created.