5.7.1 Creating and using scopes
The webHost service uses  scope

"Why do I need a scope at all? Why not simply request the service from the service provider directly? The answer is that the webhost service provider is the root application container. As a result, this container and all objects it creates, will live for the life of the application. The UserManager and the RoleManager are scoped services and depend on a scoped DbContext. I were to resolve this directly from the root container, the instances would be essentially created as singletons and outlive their intended lifetime"