6.4.2 Applying the Decorator Pattern using Scrutor
"The decorator pattern lets up wrap functionality of a service with additional functionality using another service, which conforms to a matching interface. This is a nice way to add features and extend functionality without modifying an existing class."

In this demo, the weather service is decorated with caching to reduce the number of calls made to the API. Caching could've been implemented in the service itself; however, the use of the decorator allows to keep the concern of calling the API independent of caching.